Running \(\chi^2\)-IG hierarchical model meta-analyses (LOOCV)

This function performs a meta-analysis on movement tracking data, for mean home range area (AKDE) or continuous-time speed and distance (CTSD) estimates for a sampled population. It leverages the `ctmm` R package, specifically the `meta()` function, to obtain population-level mean parameters. This function helps to assess outputs via leave-one-out cross-validation (LOOCV).

Usage

run_meta_loocv(
  rv,
  set_target = c("hr", "ctsd"),
  subpop = FALSE,
  trace = FALSE,
  .only_max_m = TRUE,
  .lists = NULL
)

Arguments

rv

A list containing outputs, settings and data objects. Must not be NULL.

set_target

Character. Research target: `"hr"` for home range or `"ctsd"` for speed & distance.

subpop

Logical. If TRUE, will run meta-analyses with groups. Default is FALSE.

trace

Logical. If TRUE, prints progress messages. Default is FALSE.

.only_max_m

Logical. If TRUE, will only run the maximum number of individuals. Default is FALSE.

.lists

A list containing already created meta inputs. Default is NULL.

Value

A data frame containing meta-analysis outputs, including estimates, errors, confidence intervals, and group information.

Author

Inês Silva i.simoes-silva@hzdr.de

Examples

if (FALSE) { # \dontrun{
# Running:
run_meta_loocv(rv, set_target = "hr")
} # }